small error

+++ multitask_classifier.py
@@ -353,7 +353,7 @@ def train_multitask(args):
                 ### similarity ----------------------------------------------
                 batch = next(sts_iterator)
                 (b_ids1, b_mask1, b_ids2, b_mask2, b_labels) = (batch['token_ids_1'], batch['attention_mask_1'], batch['token_ids_2'], batch['attention_mask_2'], batch['labels'])
-                b_ids1, b_mask1, b_ids2, b_mask2, b_labels = b_ids1.to(torch.float32).to(device), b_mask1.to(torch.float32).to(device), b_ids2.to(torch.float32).to(device), b_mask2.to(torch.float32).to(device), b_labels.to(torch.float32).to(device)
+                b_ids1, b_mask1, b_ids2, b_mask2, b_labels = b_ids1.to(device), b_mask1.to(device), b_ids2.to(device), b_mask2.to(device), b_labels.to(device)
 
                 optimizer.zero_grad()
                 if args.rlayer:
